move test files to match namespaces
