+++ tracer/test/Datadog.Trace.Tests/Agent/ApiTests.cs
@@ -17,7 +17,7 @@ using FluentAssertions;
 using Moq;
 using Xunit;
 
-namespace Datadog.Trace.Tests
+namespace Datadog.Trace.Tests.Agent
 {
     public class ApiTests
     {
